add the most basic wikidata search endpoint
