+++ inference/app/src/wikidata/__init__.py
@@ -1,0 +1,4 @@
+from ..http import fetch_url_json
+from .exact import get_wikidata_data
+from .utils import (find_alt_source_id_in_wikidata, loc_id_to_wikidata_id,
+                    mesh_id_to_wikidata_id, wikidata_id_to_alt_source_ids)
