+++ inference/app/src/wikidata/exact.py
@@ -3,26 +3,10 @@ import logging
 import re
 import time
 
-from .http import fetch_url_json
+from . import fetch_url_json
+from .utils import get_wikidata_api_response
 
 log = logging.getLogger(__name__)
-
-
-async def get_wikidata_api_response(wikidata_id):
-    if not re.match("Q[0-9]+", wikidata_id):
-        raise ValueError(f"{wikidata_id} is not a valid wikidata ID")
-
-    url = f"http://www.wikidata.org/wiki/Special:EntityData/{wikidata_id}.json"
-    response = await fetch_url_json(url)
-    if response["object"].status == 200:
-        pass
-    elif response["object"].status in [400, 404]:
-        raise ValueError(f"{wikidata_id} is not a valid wikidata ID")
-    else:
-        raise ValueError(
-            f"something unexpected happened when calling url: {url}"
-        )
-    return response["json"]["entities"][wikidata_id]
 
 
 def get_label(api_response):
